requirements.txt
async-generator==1.10
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.1
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
webdriver-manager==3.8.6
wsproto==1.2.0
